requirements.txt
Flask
Flask-Cors
azure-storage-blob
tensorflow
joblib
numpy